Fix all warnings, turn on warnings as errors (#1062)
Fix all warnings and turn on warnings as errors so they stop accumulating.
Add Mock constructor and destructor implementations to a cpp to speed up the build. Saves one minute.
Closes #1042

+++ trview.app/Elements/Entity.cpp
@@ -73,7 +73,7 @@ namespace trview
     {
     }
 
-    Entity::Entity(const IMesh::Source& mesh_source, const IMeshStorage& mesh_storage, const trlevel::ILevel& level, uint32_t room, uint32_t index, uint32_t type_id,
+    Entity::Entity(const IMesh::Source& mesh_source, const IMeshStorage& mesh_storage, const trlevel::ILevel& level, uint16_t room, uint32_t index, uint16_t type_id,
         const Vector3& position, int32_t angle, int32_t ocb, bool is_pickup)
         : _room(room), _index(index)
     {
