+++ trview.app/Elements/TypeNameLookup.h
@@ -10,7 +10,7 @@ namespace trview
     public:
         explicit TypeNameLookup(const std::string& type_name_json);
         virtual ~TypeNameLookup() = default;
-        virtual std::string lookup_type_name(trlevel::LevelVersion level_version, uint32_t type_id, uint32_t flags) const override;
+        virtual std::string lookup_type_name(trlevel::LevelVersion level_version, uint32_t type_id, uint16_t flags) const override;
         virtual bool is_pickup(trlevel::LevelVersion level_version, uint32_t type_id) const override;
     private:
         struct Type
